Move AuRaContext and AuRaState
